requirements.txt
pandas
seaborn
plotly
requests
scikit-learn
mysql-connector-python
xgboost
matplotlib
sort-dataframeby-monthorweek
sorted_months_weekdays
jinja2
fastapi
uvicorn[standard]
gunicorn
databases
sqlalchemy
pyodbc
python-multipart
azure.functions


beautifulsoup4